refactor: fully separate and conceptualize world-scale and screen-scale
feat: use matrix for scaling

+++ Components/PhysicsData.cs
@@ -8,10 +8,10 @@
 		public float MaximumSpeed => _maximumSpeed;
 
 		// ----- Fields -----
-		private int _maximumSpeed;
+		private float _maximumSpeed;
 		
 		// ----- Constructors -----
-		public PhysicsData(int maxSpeed) {
+		public PhysicsData(float maxSpeed) {
 			_maximumSpeed = maxSpeed;
 		}
 	}
